v5.1.6 BUG FIX
FIXED A SERIOUS BUG IN THE CALCULATION (OR BETTER THE ORDERING) OF
THE RENORMALIZATION CONSTANTS IN THE CODE.

The call to *observables_impurity*, calculating the imp observables and
other local quantities such as Z, scattering rate, etc. was BEFORE
the actual call to build_imp_gf, and so the calculation of the self-energy.
This led to an improper use of the former self-energy (still in the memory)
to evaluate the Z. For single site problems this went almost unnoticed because
sigma^{n-1} is almost identical to sigma^{n}.

However the bug affected the inequivalent sites calculations, leading to a
bad ordering of the zetas with respect to the index of the inequivalent site.
So, for example, site 1 used the self-energy of the Nth site at the previous loop
to get Z. Site 2 used the self-energy of site 1, and so on.

The solution is to exchange the order of calls to observables_impurity and buildgf_impurity
in ED_MAIN.

For the actual data the best fix is now to manually re-order the zeta when needed.
A small error is anyway unavoidable for one site (the first). At convergence this is
supposed to be negligible.