BUG:
there is a huge BUG appearing for Large Ns (total number of sites).
The size of the Fock space (2**2Ns) is larger for Ns=16 than the
largest possible integer(4) 2147483647
So at some random point in the code (see below, taken from HxVbath_imp.f90)
the system creates an element between i and j>2147483647 which ends up in a crash of the
code.

The only solution to this problem is to move the code to the NupNdw basis entirely
(which is already largely coded in the branch devel_NupNdw).

00000000000010000000000000000001
           1           1           5
           1           2           6
           1           3           7
           2           1           8
           2           2           9
           2           3          10
           3           1          11
           3           2          12
           3           3          13
           4           1          14
                              49                          524289                       536870913                             209                      2147483647
           4           2          15
                              49                          524289                      1073741825                             225                      2147483647
           4           3          16
                              49                          524289                     -2147483647                               0                      2147483647
-------------------------------------------------------------------------------------------^