Intermediate commit. Branch closed.
This development line will end in a separate code.
It is not possible at the current state to integrate the Nup Ndw development
into the code without messing up with the structure of the code and of the
data structures.

In particular: the states of the H should be writable as direct products very much
as you do for the map (M_up [3] TIMES M_dw [3] --> M_sector [9]). This
affects the execution a lot but also requires one to store the vectors in a different way
which is not currently supported.

Some of the updates done here can however be integrated in the master line.

CIAO