A bug for NORMAL-REPLICA has been fixed in ED_GREENS_FUNCTION.
Essentially an if-condition [if(mask)cycle] was missing, causing the
system to reduce algebraically off-diagonal elements of the GF.
If G_{ab} was not evaluated this led to wrong results for the imp. GF and
eventually to wrong Self-energy. Fixed adding back the if-condition.

Results are now better but still a bit puzzling.
Tested to the Bethe 2 degenerate bands with Hubbard only interaction.
To be tested Kanamori and/or non-degenerate situation..

With analyzed carefully the chi**2 fit and found that in some cases
the gradient is very near or exactly zero. This makes it impossible
for the bracketing of the minima subroutine in CG-NR (mnbrak) to
complete its job. After a transient the interval bracket reaches
NaN and stop the run (this is in the updated SciFortran).

We added a parameter CG_DF_EPS to set the small number h in the
gradient evaluation thru incremental ratio. Sometimes setting a
value larger than 0d0 (default) helps the fit.

In the Mott regime (U=4 Bethe 2 bands) the code does not converge anyway.