Intermediate Commit: UPDATE
This version has a new MPI implementation for the Matrix Vector product.

The sparse matrix is subdivided into MpiChunks of rows and assigned to
each thread. The corresponding rectangular matrix is further divided into
two parts: Local and Non-Local.
The Local part contains all columns such that j<row \in MpiChunks, so it
is square and small.
The Non-Local part contains the rest of the matrix and is rectangular and
possibly split into two sides (left and right of the local part).

This enables to perform the following steps in the MPI matrix-vector product:
1. The local Matrix times local vector chunk is performed with no restriction
at each thread. This provides a first (and possibly substantial) contribution
to the overall result and does not require any communication among threads.

2. The non-local matrix contribution is evaluated as follows:
 A. The range of non-zero elements columns is determined analysing the
 sparse matrix. If direct mat-vec product is required a *dryrun* is performed
 to store only this info. This is a small price to pay (1/#calls of mat-vec).
 B. Using the above range the corresponding range of threads having the necessary
 rows required to perform the mat-vec product is determined per each thread.
 Using this info, each thread GATHERV the chunks of vector needed for the product
 and corresponding to the non-zero elements in the matrix only.
 Depending on the sparseness of the matrix this saves a lot of communication:
 essentially #columns_in_the_range/#dimension_of_the_vector
 which can be less than 10% and typically increases with Ncpu.

Note that this optimization does not require the Nup_Ndw basis setup, which is
held separate so far and provides a further improvement.

Ready to test this implementation for the Bethe lattice.