Intermediate Committ:
Tested for the Bethe Lattice MPI.

Unfortunately the distribution of the columns is very spread, which
makes the limitation to the columns range almost inessential.

We need further checks for larger systems.
I wonder if one can find a way to construct the states so that
elements are far less spread...

In any case the Gatherv calls are not working when limited to Imin:Imax.
There is a mismatch between send_buffer > rec_buffer which I could not
track down so far.

So I decided to fall back on the solution without analysis of the
columns distribution, which remains in the code but commented.
The solution in this case seems correct in the  serial and parallel cases.
More testing are needed against the most recent version of the Bethe MPI
code from the master branch.

The Direct Mat-Vec product has a lower performance in any case,
because we can not distinguish a priori the local from the non
local contribution.

Waiting for further testing in the near future. Not ready to merge.