v5.1.7 UPDATE
Updated code.
The Greens functions module has been splitted into many modules,
one for each type of green's function (normal, superc, nonsu2) or
susceptility to calculate.
This update follows closely the one implemented for the NEW code
with NupNdw structure which is being developed separately. The aim
is to have, possibly but not necessarily, a single code.