v5.2.0 UPDATE + COMPILATION BREAKDOWN FIX
This version features a new MPI structure, mostly consequence
of the updates introduced in the MPI interface in Scifor, which
in turn broken down the compilation.

The essential updates are the followin:
- the MPI split of the arrays (vector and H matrices) are performed
in the code by the user. The SciFor procedures accept on both input
and output "already split" arrays.
- the vectors are stored in chunks (split) on the corresponding
threads (memory gain). Only the master gather the chunks and performed
the required operations, before scatter it back to the threads for
calculation (i.e. H*v in parallel).

Tests are done for the BHZ model and Attractive HM. Perfect results
within Lancozs error accuracy.