v5.3.0 UPDATE
The code has been updated with a novel data structure for the
sparse matrix storage.
This structure features a O(1) access to the elements, yet it
stores only the non-zero elements using a home-brew dynamic
array, i.e. a growing array. The matrix is defined as an
"array-of-dynamic arrays".

At the price of a larger use of memory this approach
offers a further improvement in the speed-up of the code.