Intermediate commit
Fixing the state_list....restart files in ED_SETUP.
This was different from what is actually printed in ED_DIAG after
calculation.

There is still a part to be fixed in ED_DIAG for the T=0 case.
This IO operation is different between this and the new code.
Recall it if you are passing solutions from here to there.