v5.5.5 UPDATE & (possible) BUG FIX
- ED_SETUP: we changed the line reading the state_list.restart file to
a default format (read(unit,*,iostat=status)). Using a specific format
was somehow returnin incorrect quantum numbers and sector labels.
Now it should work.

- ED_GF_NORMAL: we removed (again!!!) the spin symmetrizing part
in the beginning of the code, having ed_para=T and Nspin=2. This
SHOULD NOT BE TURNED BACK. It is just wrong and misleading for the
users.
If someone needs to symmetrize should do that at the exit of the
calculation or symmetrize the bath. Otherwise we can cause
problems to others actually doing magnetic calculation.
ed_para must be removed or set to .false. as default in the future.

code tested on ed_nano_isoc_test. old results reproduced perfectly.