5.6.0beta CRITICAL UPDATE: BUG FIX
A bug has been possibly eliminated in the construction of the H matrix.
We realized that H was constructed transposed with respect to the expected
normal form. In the case of real components only H is symmetric so the bug
was harmless. VIceversa in the case of cmplx components (e.g. in the nonSU2
case) the transposition produces possible wrong results for the GF (same
spectrum but wrong symmetry of the GF), as the matrix requires a cmplx
conjugation.

Fixed by exchanging everywhere rows and cols indices (this keep
readibility of the code).

Test for BHZ and Attractive HM are OK.
More tests to do for the
+direct HxV channel
+nonSU2 ed_mode

Consider this a beta release.