Intermediate commit.
Fixing a bug appearing since the latest update concerning the
transposed matrix, wrong hermiticity, in the MPI channel.

The MPI sparse matrices are distributed along rows. Now the
lastest change introduced to fix the transposed construction of the Hs
make us rolling along columns. I naively thought that this would have
been the same for the MPI, but obviously it is not.
The reason I suspect is that application of operators can produce
highly non-local elements in the distrubuted matrix.
In poor words: if you start from any column j, the processor X can get
a row i which is *not* in its own part of the distributed matrix.
TO ensure that we should distribute along columns. But this is bad
because matrix-vector produxt proceeds along rows...

To fix the issue I know conjg all the complex non-diagonal entries.
This should correspond to build H^+ rather than H. Hermiticity
guarantees that the two are the same stuff so...

It apparently works for the simplest Bethe 3bands case at 1 loop.