Included and tested routine to rotate the interaction.
Kanamori has found to be machine precision invariant against SO(3) real roration
of the orbital subspace, as expected.
The effect of the rotation that brings from the cubic harmonics basis
to the spherical harmonics basis, SU(3), produces a substantial change
in the tensor elements. The new form of the interaction is identical
to the one analytically predicted (see notes).