Sneaky bug
It seems like intel fortran complier (v18.0.1) does not like to have
unallocated variables passed to subroutines optional arguments.
"forrtl: severe (408): fort: (7): Attempt to use pointer V when it is not associated with a target"
was given when calling MPI_Scatterv inside the scatter_vector_MPI subroutine.
This is called when the code performs the standard lanczos for the the Lehmann Green's
function via lanc_build_gf_nonsu2_*Orb_*Spin_c
The V pointer refers to the vvinit vector which is built on the master node and then scattered among threads.
The code was allocating vvinit *only* for master which explains the error.
This is not a problem using gfortran (I tested it) and was already categorized as ifort issue:
https://software.intel.com/en-us/forums/intel-fortran-compiler/topic/557015
but apparently not fixed (the -check all flag for us is active only in debug mode).
I fixed just allocating vvinit for all threads in the ED_GF_NONSU2 module I am using now
but probably this is going to be an issue for any ifort lover.
I would suggest to modify also the other two Gf modules if a merge is planned.

God bless Steve Lionel the almighty (which is still refusing to accept me within his linkedin bootlicker).